requirements_api.txt
# JWT Authentication and REST API Requirements
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
